- Shuffling Added - All UI placed - Success Game Ending Improved

+++ Assets/Scripts/LevelProperties/LevelRequirementTime.cs
@@ -4,5 +4,5 @@ using UnityEngine;
 
 public class LevelRequirementTime : MonoBehaviour
 {
-    public int RequirementTime;
+    public float RequirementTime;
 }
